Fix playback not working on some systems/configurations.
The player was waiting for a quarter second of output buffer space to become available before sending any samples. If the configured playback buffer size on a system was smaller than that, samples would never be sent. Instead, send any available samples as soon as the output component is ready to accept them.