Fix wrong ELF OS ABI in FreeBSD aarch64 AppImage runtime.
Apparently, there is an issue with GNU ld shipped with FreeBSD aarch64 where it sets the ELF OS ABI to GNU when IFUNCs are present in a binary (which is the case here with symbols imported from libmd.a). This is the correct behavior on Linux, but not on FreeBSD. Just force the OS ABI to FreeBSD here which also supports IFUNCs.